src/requirements.txt
Django==3.2
celery==5.1.2
psycopg2-binary==2.8.6
django-environ==0.4.5
aiogram==2.13
python-dotenv==0.17.0
SQLAlchemy==1.4.12
asyncpg==0.23.0
aioredis==1.3.1
requests==2.25.1
gunicorn==20.1.0
sentry-sdk==1.1.0
Jinja2==3.0.0
Pillow==8.4.0
django-bootstrap4==3.0.1
click==7.1.2
loguru==0.5.3
redis==3.5.3
sqlalchemy-json==0.4.0
gevent==21.1.2
django-flat-json-widget==0.1.3
openpyxl==3.0.7
django-map-widgets==0.3.2
GeoAlchemy2==0.9.4
lxml==4.6.3
uvloop==0.16.0
ujson==4.2.0
aiohttp[speedups]
qrcode[pil]